merges subclass config for pandas data frame
